release 0.1.7 source code for go

+++ services/cdn/v1/model/model_urls.go
@@ -25,10 +25,10 @@ type Urls struct {
 	TaskId *int64 `json:"task_id,omitempty"`
 
 	// 修改时间戳（毫秒）
-	ModifyTime *int32 `json:"modify_time,omitempty"`
+	ModifyTime *int64 `json:"modify_time,omitempty"`
 
 	// 创建时间戳（毫秒）
-	CreateTime *int32 `json:"create_time,omitempty"`
+	CreateTime *int64 `json:"create_time,omitempty"`
 
 	// 文件类型，目录还是文件
 	FileType *string `json:"file_type,omitempty"`
